# api_server/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: info@payloadops.com
    name: Payload Ops Team
  description: "The Payload Ops API streamlines management of AI prompts, projects,\
    \ organizations, teams, and users through conventional HTTP requests. This platform\
    \ enables efficient automation and control of resources, providing robust tools\
    \ for developers to manage settings, memberships, and activities seamlessly."
  title: Payload Ops API
  version: "1.0"
servers:
- description: Production server (uses live data)
  url: https://api.payloadops.com/v1
- description: Sandbox server (uses test data)
  url: https://sandbox-api.payloadops.com/v1
security:
- ApiKeyAuth: []
- BearerAuth: []
paths:
...
  /organizations/{orgId}/teams:
    get:
      description: |
        Retrieves a comprehensive list of all teams within a specified organization. This endpoint is essential for administrators and organizational leaders to view and manage team structures, understand team compositions, and plan resource allocation effectively.

        **Key Features**:
        - **Comprehensive Overview**: Provides detailed views of all teams, including their names, descriptions, and member counts.
        - **Management and Planning**: Supports strategic planning and management by offering insights into team distributions and roles within the organization.

        **Use Cases**:
        - Executives and managers reviewing team structures for strategic planning.
        - HR departments managing resource allocations and internal audits of team distributions.
      operationId: listTeams
      parameters:
      - description: The identifier of the organization whose teams are being listed.
        explode: false
        in: path
        name: orgId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Team'
                type: array
          description: "Successfully retrieved a list of all teams within the organization,\
            \ each entry containing details such as team ID, name, and member list."
        "500":
          description: "A server error occurred, preventing the retrieval of team\
            \ information."
      security:
      - BearerAuth: []
      summary: List all teams in an organization
      tags:
      - Teams
    post:
      description: |
        Creates a new team within an organization based on the provided details such as team name, description, and intended members. This operation facilitates the expansion and structuring of an organization's teams to meet evolving business needs and project demands.

        **Key Features**:
        - **Team Formation**: Enables the formation of new teams tailored to specific projects or organizational functions.
        - **Customizable Setup**: Allows for detailed definition of team characteristics including leadership roles and member responsibilities.

        **Use Cases**:
        - Forming new project teams as part of business expansion or new project initiation.
        - Restructuring existing departments to enhance efficiency or focus on new strategic areas.
      operationId: createTeam
      parameters:
      - description: The identifier of the organization where the new team will be
          created.
        explode: false
        in: path
        name: orgId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamInput'
        description: "Contains the necessary attributes for creating a new team, including\
          \ name, description, and initial members if applicable."
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: "The team was created successfully and is now part of the organization,\
            \ with detailed information about the new team provided."
        "400":
          description: "Invalid input, typically due to missing required fields or\
            \ improperly formatted data."
        "500":
          description: "A server error occurred, preventing the creation of the new\
            \ team."
      security:
      - BearerAuth: []
      summary: Create a new team for an organization
      tags:
      - Teams
  /organizations/{orgId}/teams/{teamId}:
    delete:
      description: |
        Deletes a specific team from an organization based on its team ID. This operation is critical for managing organizational structure, particularly when a team is dissolved or merged with another.

        **Key Features**:
        - **Streamlined Organizational Structure**: Helps maintain a clean and efficient organizational hierarchy by removing outdated or unnecessary teams.
        - **Security and Compliance**: Ensures that all related data is securely handled and removed in compliance with data protection policies.

        **Use Cases**:
        - Removing teams that are no longer active or have been restructured.
        - Ensuring that team data does not persist in the system unnecessarily, particularly in sensitive environments.
      operationId: deleteTeam
      parameters:
      - description: The identifier of the organization from which the team will be
          deleted.
        explode: false
        in: path
        name: orgId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the team to be deleted.
        explode: false
        in: path
        name: teamId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: "Team deleted successfully, with no remaining data stored."
        "404":
          description: Either the specified team or the organization was not found.
        "500":
          description: "A server error occurred, preventing the deletion of the team."
      security:
      - BearerAuth: []
      summary: Delete a specific team
      tags:
      - Teams
    get:
      description: |
        Retrieves detailed information about a specific team within an organization, using its unique team ID. This operation is crucial for administrators who need to manage team settings, understand team composition, and review team-specific roles and permissions.

        **Key Features**:
        - **Comprehensive Team Details**: Provides complete information about the team, including its members, roles, projects, and operational statistics.
        - **Management and Oversight**: Essential for overseeing team configurations and ensuring that team setups align with organizational policies and goals.

        **Use Cases**:
        - Team leaders reviewing team composition and roles for project assignments.
        - HR and management accessing team details for planning and resource allocation.
      operationId: getTeam
      parameters:
      - description: The identifier of the organization to which the team belongs.
        explode: false
        in: path
        name: orgId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the team to retrieve.
        explode: false
        in: path
        name: teamId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: Successfully retrieved detailed information about the team.
        "404":
          description: The specified team or organization was not found.
        "500":
          description: "A server error occurred, preventing the retrieval of the team."
      security:
      - BearerAuth: []
      summary: Get a specific team by ID
      tags:
      - Teams
    put:
      description: |
        Updates the details of a specific team within an organization. This endpoint allows for the modification of team data such as name, description, and member roles. It supports dynamic team management and ensures that team information remains current and relevant to organizational needs.

        **Key Features**:
        - **Dynamic Team Management**: Enables changes to team structure, roles, and project assignments.
        - **Adaptable to Changing Needs**: Facilitates quick updates to team details in response to project shifts or strategic realignments.

        **Use Cases**:
        - Administrators adjusting team roles and permissions to reflect new project requirements.
        - Team leaders modifying team descriptions or details to better define the team’s purpose or objectives.
      operationId: updateTeam
      parameters:
      - description: The identifier of the organization where the team is registered.
        explode: false
        in: path
        name: orgId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the team to be updated.
        explode: false
        in: path
        name: teamId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamInput'
        description: Contains the updated information for the team.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
          description: "Team information updated successfully, reflecting new settings\
            \ and configurations."
        "400":
          description: "Invalid input, such as missing required data or unsupported\
            \ modifications."
        "404":
          description: Either the specified team or the organization was not found.
        "500":
          description: "A server error occurred, preventing the update of the team."
      security:
      - BearerAuth: []
      summary: Update a specific team
      tags:
      - Teams
...
  /projects/{projectId}/prompts:
    get:
      description: |
        Retrieves a comprehensive list of all prompts associated with a specific project. This operation is essential for users to view and manage all the prompts under a project, facilitating quick access to any particular prompt and understanding the scope of the project's content.

        **Key Features**:
        - **Comprehensive Overview**: Provides a complete list of all prompts, helping in the overall management and navigation of project content.
        - **Efficient Project Management**: Enables project managers and team members to quickly assess the progress and status of various prompts within the project.

        **Use Cases**:
        - Developers and content creators looking to overview or navigate through all the prompts associated with a specific project.
        - Useful in scenarios where multiple teams or stakeholders need to review or access prompt information collectively.
      operationId: listPrompts
      parameters:
      - description: The unique identifier of the project whose prompts are to be
          listed.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Prompt'
                type: array
          description: Successfully retrieved a list of all prompts in the specified
            project.
        "404":
          description: "Project not found, indicating the specified project ID does\
            \ not exist."
        "500":
          description: "A server error occurred, preventing the retrieval of prompts."
      summary: List all prompts in a project
      tags:
      - Prompts
    post:
      description: |
        Creates a new prompt within a specified project. This endpoint is crucial for expanding project content by adding new prompts, allowing users to input initial details for a new prompt which then becomes part of the project's documentation or content offerings.

        **Key Features**:
        - **Content Expansion**: Allows for the easy addition of new content to a project, supporting dynamic growth and updates.
        - **Customizable Input**: Users can define various attributes of the prompt such as title, description, and specific content parameters through the `PromptInput` schema.

        **Use Cases**:
        - Content managers and creators adding new thematic prompts or content pieces to existing projects.
        - Automating content creation processes by integrating with content management systems that can trigger prompt creations through API calls.
      operationId: createPrompt
      parameters:
      - description: The unique identifier of the project within which the new prompt
          is to be created.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptInput'
        description: "Contains the necessary attributes for creating a new prompt,\
          \ such as title, description, and any specific content elements."
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prompt'
          description: The prompt was created successfully and is now part of the
            project.
        "400":
          description: "Invalid input provided in the request body, such as missing\
            \ required fields or data that does not conform to the expected format."
        "404":
          description: "Project not found, indicating the specified project ID does\
            \ not exist."
        "500":
          description: "A server error occurred, preventing the creation of the new\
            \ prompt."
      summary: Create a new prompt in a project
      tags:
      - Prompts
  /prompts/{promptId}:
    delete:
      description: |
        Deletes a specific prompt from a project, removing it permanently from the system. This operation is used when a prompt is no longer needed or relevant, or as part of content cleanup and management strategies.

        **Key Usage**:
        - Removing outdated or irrelevant prompts to maintain a clean and efficient project environment.
        - Complying with data retention policies by deleting prompts that are no longer necessary.
      operationId: deletePrompt
      parameters:
      - description: The identifier of the project from which the prompt will be deleted.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the prompt to be deleted.
        explode: false
        in: path
        name: promptId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: The prompt was deleted successfully; no content is returned.
        "404":
          description: Either the prompt or the project was not found.
        "500":
          description: "A server error occurred, preventing the deletion of the prompt."
      summary: Delete a specific prompt from a project
      tags:
      - Prompts
    get:
      description: |
        Retrieves detailed information about a specific prompt within a project, including its content, metadata, and any associated versioning details. This endpoint is essential for users who need to view the current state of a prompt or to fetch its details for further processing or display.

        **Key Usage**:
        - Viewing the complete details of a prompt for editorial or review purposes.
        - Fetching prompt data for integration with other tools or services.
      operationId: getPrompt
      parameters:
      - description: The identifier of the project to which the prompt belongs.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the prompt being retrieved.
        explode: false
        in: path
        name: promptId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prompt'
          description: Successfully retrieved detailed information about the prompt.
        "404":
          description: Either the prompt or the project was not found.
        "500":
          description: "A server error occurred, preventing the retrieval of the prompt."
      summary: Retrieve a specific prompt within a project
      tags:
      - Prompts
    put:
      description: |
        Updates the content or metadata of a specific prompt within a project. This operation allows users to modify existing prompts to reflect changes in requirements, updates, or corrections. It is crucial for maintaining the accuracy and relevancy of prompt content.

        **Key Usage**:
        - Modifying the text or other properties of a prompt as part of content management workflows.
        - Updating prompt settings or configurations based on new inputs or feedback.
      operationId: updatePrompt
      parameters:
      - description: The identifier of the project that contains the prompt.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the prompt being updated.
        explode: false
        in: path
        name: promptId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptInput'
        description: JSON payload containing the updated data for the prompt.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prompt'
          description: The prompt was updated successfully.
        "400":
          description: "Invalid input, such as malformed data or incomplete required\
            \ fields."
        "404":
          description: Either the prompt or the project was not found.
        "500":
          description: "A server error occurred, preventing the update of the prompt."
      summary: Update a specific prompt in a project
      tags:
      - Prompts
...
components:
  schemas:
...
    Team:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        description: description
        id: id
        orgId: orgId
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Unique identifier for the team
          type: string
        name:
          description: Name of the team
          type: string
        description:
          description: A brief description of the team
          type: string
        orgId:
          description: The organization to which the team belongs
          type: string
        createdAt:
          description: Timestamp of team creation
          format: date-time
          type: string
        updatedAt:
          description: Timestamp of the last update to the team
          format: date-time
          type: string
      type: object
    TeamInput:
      example:
        name: name
        description: description
      properties:
        name:
          description: Name of the team
          type: string
        description:
          description: A brief description of the team
          type: string
      required:
      - name
      type: object
...
    Prompt:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        description: description
        id: id
        projectId: projectId
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Unique identifier for the prompt
          type: string
        name:
          description: Name of the prompt
          type: string
        description:
          description: A brief description of the project
          type: string
        projectId:
          description: Identifier of the project this prompt belongs to
          type: string
        createdAt:
          description: Timestamp when the prompt was created
          format: date-time
          type: string
        updatedAt:
          description: Timestamp when the prompt was last updated
          format: date-time
          type: string
      type: object
    PromptInput:
      example:
        name: name
        description: description
      properties:
        name:
          description: Name of the prompt
          type: string
        description:
          description: A brief description of the project
          type: string
      required:
      - name
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: |
        API Key authentication provides a straightforward method for clients to authenticate requests. This scheme expects an API key to be supplied in the headers of each request. Clients must include the API key in the 'X-API-KEY' header field.

        **How to use**:
        - Obtain an API key from the API provider.
        - Include the API key in the header of each request to the API as follows:
          ```
          X-API-KEY: {api_key}
          ```
        - The API validates the key against its database, and if the key is valid, allows access to the requested resources.

        **Advantages**:
        - Simple and easy to implement.
        - Suitable for accessing public APIs where security requirements are not excessively strict.

        **Limitations**:
        - Less secure than more complex authentication methods, such as tokens, as API keys can be easily exposed in transit or storage.
        - API keys do not expire by default and revoking them requires direct intervention.
      in: header
      name: X-API-KEY
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      description: |
        Bearer Token authentication uses JSON Web Tokens (JWT) to verify the identity of the requester. This scheme is included in the Authorization header of each HTTP request using the Bearer format.

        **How to use**:
        - Obtain a JWT from the authentication server. Typically, this involves an initial request with credentials (username/password) to receive the token.
        - Include the token in the Authorization header of your HTTP requests as follows:
          ```
          Authorization: Bearer {token}
          ```
        - The server decodes the JWT to verify its validity and authorizes the request based on the token's payload and signature.

        **Advantages**:
        - Offers more security than API key authentication as the token content is protected by cryptographic algorithms.
        - Tokens can be designed to expire after a certain period, enhancing security by limiting the duration of their validity.

        **Limitations**:
...