# app_1/pkg/handlers/prompt/createprompt.go
package prompthandler

import (
	"encoding/json"
	"net/http"
	"plato/app/pkg/model"
	"plato/app/pkg/util"
	promptservicemodel "plato/app_1/pkg/model/prompt/service"
	promptservice "plato/app_1/pkg/service/prompt"
	"strings"

	"github.com/go-chi/render"
)

func CreatePromptHandler(w http.ResponseWriter, r *http.Request) {
	promptService, _ := promptservice.NewService()
	validator := util.GetValidator()

	var createPromptRequest promptservicemodel.CreatePromptRequest
	if err := json.NewDecoder(r.Body).Decode(&createPromptRequest); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	if err := validator.Struct(createPromptRequest); err != nil {
		render.Render(w, r, model.ErrorResponseRenderer(http.StatusBadRequest, err.Error()))
		return
	}

	if len(createPromptRequest.Branch) == 0 {
		createPromptRequest.Branch = "main"
	}

	urlSlices := strings.Split(r.URL.Path, "/")
	projectId := urlSlices[4]

	response, err := promptService.CreatePrompt(
		r.Context(),
		projectId,
		createPromptRequest,
	)

	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	render.Status(r, http.StatusOK)
	render.JSON(w, r, response)
}